feat: add isJson method to CurlHttpClient class to improve handling of response content type
The `isJson` method is added to the `CurlHttpClient` class. This method checks if the provided content type is either 'application/json' or 'application/hal+json'. This method is used in the `handleResponse` method to determine if the response body should be parsed as JSON. This improves the handling of different content types in the client.